Basically rewrote the entire plugin
This commit introduces major changes and sets the sail for a fork that is really different from the main plugin. Currently the main difference is that we do not make use of Recpies, the cost system has been fully reworked and that major refractors have been performed on how the plugin handles materials, now it uses EnumMaps instead of String HashMaps, which should make the code run in a more efficent manner. It also introduces the first documentations and the addition of an Utillity class. Now, the plugin in it's current state hasn't been tested as of now and will have some issues, it also won't be able to export the cost configuration files correctly as there are no cost configuration files in the resource folder, converting the old system into the new system shouldn't take too long. The biggest issue will be runtime however as the data is saved in a single big YAML file, which in addition with snakeyaml is a really stupid way of saving data. I think it's better to save the data one day via a price history file or something similar that is saved in binary, which saves the need of having to write to the configuration every time something was changed. Additionally it should be faster as as of now it might take a few dozens of milliseconds to save a single price change within a 1000 lines configuration file.

Signed-off-by: Geolykt <admin@geolykt.de>